+++ packages/renderers-js/e2e/dummy/test/_setup.ts
@@ -1,11 +1,12 @@
 import {
+  type BaseTransactionMessage,
   type Commitment,
-  type CompilableTransactionMessage,
   type Rpc,
   type RpcSubscriptions,
   type SolanaRpcApi,
   type SolanaRpcSubscriptionsApi,
   type TransactionMessageWithBlockhashLifetime,
+  type TransactionMessageWithFeePayer,
   type TransactionSigner,
   airdropFactory,
   createSolanaRpc,
@@ -48,7 +49,11 @@ export const generateKeyPairSignerWithSol = async (
 export const createDefaultTransaction = async (
   client: Client,
   feePayer: TransactionSigner
-) => {
+): Promise<
+  BaseTransactionMessage &
+    TransactionMessageWithFeePayer &
+    TransactionMessageWithBlockhashLifetime
+> => {
   const { value: latestBlockhash } = await client.rpc
     .getLatestBlockhash()
     .send();
@@ -61,7 +66,8 @@ export const createDefaultTransaction = async (
 
 export const signAndSendTransaction = async (
   client: Client,
-  transactionMessage: CompilableTransactionMessage &
+  transactionMessage: BaseTransactionMessage &
+    TransactionMessageWithFeePayer &
     TransactionMessageWithBlockhashLifetime,
   commitment: Commitment = 'confirmed'
 ) => {