+++ packages/renderers-js/src/fragments/instructionData.ts
@@ -1,4 +1,4 @@
-import { InstructionNode } from '@codama/nodes';
+import { InstructionNode, structTypeNodeFromInstructionArgumentNodes } from '@codama/nodes';
 import { getLastNodeFromPath, NodePath } from '@codama/visitors-core';
 
 import type { GlobalFragmentScope } from '../getRenderMapVisitor';
@@ -10,6 +10,7 @@ export function getInstructionDataFragment(
     scope: Pick<GlobalFragmentScope, 'customInstructionData' | 'nameApi'> & {
         dataArgsManifest: TypeManifest;
         instructionPath: NodePath<InstructionNode>;
+        size: number | null;
     },
 ): Fragment {
     const { instructionPath, dataArgsManifest, nameApi, customInstructionData } = scope;
@@ -23,5 +24,7 @@ export function getInstructionDataFragment(
         manifest: dataArgsManifest,
         name: instructionDataName,
         nameApi,
+        node: structTypeNodeFromInstructionArgumentNodes(instructionNode.arguments),
+        size: scope.size,
     });
 }