+++ packages/renderers-js/src/fragments/instructionParseFunction.ts
@@ -49,7 +49,8 @@ export function getInstructionParseFunctionFragment(
     })
         .mergeImportsWith(dataTypeFragment)
         .addImports('generatedPrograms', [programAddressConstant])
-        .addImports('solanaInstructions', ['type IInstruction'])
-        .addImports('solanaInstructions', hasAccounts ? ['type IInstructionWithAccounts', 'type IAccountMeta'] : [])
-        .addImports('solanaInstructions', hasData ? ['type IInstructionWithData'] : []);
+        .addImports('solanaInstructions', ['type Instruction'])
+        .addImports('solanaInstructions', hasAccounts ? ['type InstructionWithAccounts', 'type AccountMeta'] : [])
+        .addImports('solanaInstructions', hasData ? ['type InstructionWithData'] : [])
+        .addImports('solanaCodecsCore', hasData ? ['type ReadonlyUint8Array'] : []);
 }