+++ packages/renderers-js/src/fragments/typeCodec.ts
@@ -1,3 +1,5 @@
+import type { TypeNode } from '@codama/nodes';
+
 import { GlobalFragmentScope } from '../getRenderMapVisitor';
 import { TypeManifest } from '../TypeManifest';
 import { Fragment, fragmentFromTemplate, mergeFragments } from './common';
@@ -11,22 +13,26 @@ export function getTypeCodecFragment(
         encoderDocs?: string[];
         manifest: Pick<TypeManifest, 'decoder' | 'encoder'>;
         name: string;
+        node: TypeNode;
+        size: number | null;
     },
 ): Fragment {
     const { name, manifest, nameApi } = scope;
+    const codecType = typeof scope.size === 'number' ? 'FixedSizeCodec' : 'Codec';
     return mergeFragments(
         [
             getTypeEncoderFragment({ ...scope, docs: scope.encoderDocs }),
             getTypeDecoderFragment({ ...scope, docs: scope.decoderDocs }),
             fragmentFromTemplate('typeCodec.njk', {
                 codecFunction: nameApi.codecFunction(name),
+                codecType,
                 decoderFunction: nameApi.decoderFunction(name),
                 docs: scope.codecDocs,
                 encoderFunction: nameApi.encoderFunction(name),
                 looseName: nameApi.dataArgsType(name),
                 manifest,
                 strictName: nameApi.dataType(name),
-            }).addImports('solanaCodecsCore', ['type Codec', 'combineCodec']),
+            }).addImports('solanaCodecsCore', [`type ${codecType}`, 'combineCodec']),
         ],
         renders => renders.join('\n\n'),
     );