+++ packages/renderers-js/src/fragments/typeEncoder.ts
@@ -1,3 +1,5 @@
+import { isDataEnum, isNode, TypeNode } from '@codama/nodes';
+
 import type { GlobalFragmentScope } from '../getRenderMapVisitor';
 import { TypeManifest } from '../TypeManifest';
 import { Fragment, fragmentFromTemplate } from './common';
@@ -7,16 +9,23 @@ export function getTypeEncoderFragment(
         docs?: string[];
         manifest: Pick<TypeManifest, 'encoder'>;
         name: string;
+        node: TypeNode;
+        size: number | null;
     },
 ): Fragment {
-    const { name, manifest, nameApi, docs = [] } = scope;
+    const { name, node, manifest, nameApi, docs = [] } = scope;
+    const encoderType = typeof scope.size === 'number' ? 'FixedSizeEncoder' : 'Encoder';
+    const useTypeCast = isNode(node, 'enumTypeNode') && isDataEnum(node) && typeof scope.size === 'number';
+
     return fragmentFromTemplate('typeEncoder.njk', {
         docs,
         encoderFunction: nameApi.encoderFunction(name),
+        encoderType,
         looseName: nameApi.dataArgsType(name),
         manifest,
         strictName: nameApi.dataType(name),
+        useTypeCast,
     })
         .mergeImportsWith(manifest.encoder)
-        .addImports('solanaCodecsCore', 'type Encoder');
+        .addImports('solanaCodecsCore', `type ${encoderType}`);
 }