+++ packages/errors/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @codama/errors
 
+## 1.3.1
+
+### Patch Changes
+
+- Updated dependencies []:
+    - @codama/node-types@1.3.1
+
 ## 1.3.0
 
 ### Patch Changes