+++ packages/renderers/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @codama/renderers
 
+## 1.0.28
+
+### Patch Changes
+
+- Updated dependencies [[`3194644`](https://github.com/codama-idl/codama/commit/3194644cd029a2468afb4b66096f53b69c154508)]:
+    - @codama/renderers-js@1.3.2
+    - @codama/renderers-js-umi@1.1.17
+    - @codama/renderers-rust@1.1.3
+
+## 1.0.27
+
+### Patch Changes
+
+- Updated dependencies [[`a305a91`](https://github.com/codama-idl/codama/commit/a305a91e04d2d420ac57c31ccbb0ea98fc7f8823)]:
+    - @codama/renderers-rust@1.1.2
+
+## 1.0.26
+
+### Patch Changes
+
+- Updated dependencies [[`6466226`](https://github.com/codama-idl/codama/commit/6466226471340209932a073176090b262aae0b84), [`9536f33`](https://github.com/codama-idl/codama/commit/9536f337370535a6b747a0128ef0c030ae65d299)]:
+    - @codama/renderers-js@1.3.1
+
+## 1.0.25
+
+### Patch Changes
+
+- Updated dependencies [[`855c094`](https://github.com/codama-idl/codama/commit/855c094e0a47ca0cc179bd2516f6dcfda1e3d552), [`364578e`](https://github.com/codama-idl/codama/commit/364578e45eca74fa24b452cc9ef22222da03ba39)]:
+    - @codama/renderers-js@1.3.0
+    - @codama/renderers-js-umi@1.1.16
+    - @codama/renderers-rust@1.1.1
+
 ## 1.0.24
 
 ### Patch Changes