+++ packages/validators/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @codama/validators
 
+## 1.3.1
+
+### Patch Changes
+
+- Updated dependencies [[`71caa87`](https://github.com/codama-idl/codama/commit/71caa87d7010ecfeed5fb08e79955e5a7e8c0779)]:
+    - @codama/visitors-core@1.3.1
+    - @codama/errors@1.3.1
+    - @codama/nodes@1.3.1
+
 ## 1.3.0
 
 ### Patch Changes